logs and statistics live streaming improvement, async timeout and connections per route increased, #29

+++ src/main/java/com/thecookiezen/infrastructure/docker/StatisticsLite.java
@@ -10,15 +10,15 @@ class StatisticsLite {
 
     private int rx_bytes;
     private int tx_bytes;
-    private int memoryUsage;
-    private int cpuTotalUsage;
+    private long memoryUsage;
+    private long cpuTotalUsage;
 
     public StatisticsLite(Statistics statistics) {
         Map<String, Integer> eth0 = (Map<String, Integer>) statistics.getNetworks().get("eth0");
         rx_bytes = eth0.get("rx_bytes");
         tx_bytes = eth0.get("tx_bytes");
         memoryUsage = (int) statistics.getMemoryStats().get("usage");
-        Map<String, Integer> cpu_usage = (Map<String, Integer>) statistics.getCpuStats().get("cpu_usage");
+        Map<String, Long> cpu_usage = (Map<String, Long>) statistics.getCpuStats().get("cpu_usage");
         cpuTotalUsage = cpu_usage.get("total_usage");
     }
 }
